Assignment Two
The goal of this assignment was to help understanding of data manipulation and visualisation tools for machine learning. The purpose was to implement common data handling methods on real-world observations. To validate the effectiveness of the implemented methods, we were also required to perform data analysis tasks to draw useful conclusions. In particular, the following topics were be reviewed: Cross Industry Standard Process for Data Mining (CRISP-DM), Exploratory Data Analysis (EDA), Data Preparation, Feature Manipulation